requirements.txt
streamlit==1.25.0
pandas==2.0.3
plotly==5.15.0
sqlalchemy==2.0.19
numpy==1.24.3
pyyaml==6.0.1
watchdog==3.0.0
